+++ src/main/kotlin/com/github/x0x0b/codexlauncher/actions/LaunchCodexAction.kt
@@ -2,6 +2,7 @@ package com.github.x0x0b.codexlauncher.actions
 
 import com.github.x0x0b.codexlauncher.http.HttpTriggerService
 import com.github.x0x0b.codexlauncher.settings.CodexLauncherSettings
+import com.github.x0x0b.codexlauncher.terminal.CodexTerminalManager
 import com.intellij.notification.NotificationGroupManager
 import com.intellij.notification.NotificationType
 import com.intellij.openapi.actionSystem.AnAction
@@ -11,43 +12,30 @@ import com.intellij.openapi.components.service
 import com.intellij.openapi.diagnostic.logger
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.util.Key
-import com.intellij.openapi.wm.ToolWindowManager
-import com.intellij.terminal.ui.TerminalWidget
-import com.intellij.ui.content.Content
-import org.jetbrains.plugins.terminal.TerminalToolWindowManager
-import org.jetbrains.plugins.terminal.TerminalToolWindowFactory
 
 class LaunchCodexAction : AnAction("Launch Codex", "Open a Codex terminal", null), DumbAware {
-    
+
     companion object {
         private const val CODEX_COMMAND = "codex"
         private const val NOTIFICATION_TITLE = "Codex Launcher"
-        private val CODEX_TERMINAL_KEY = Key.create<Boolean>("codex.launcher.codexTerminal")
-        private val CODEX_TERMINAL_RUNNING_KEY = Key.create<Boolean>("codex.launcher.codexTerminal.running")
-        private val CODEX_TERMINAL_CALLBACK_KEY = Key.create<Boolean>("codex.launcher.codexTerminal.callbackRegistered")
     }
-    
+
     private val logger = logger<LaunchCodexAction>()
 
-    private data class CodexTerminal(val widget: TerminalWidget, val content: Content)
     override fun actionPerformed(e: AnActionEvent) {
         val project = e.project
         if (project == null) {
             logger.warn("No project context available for Codex launch")
             return
         }
-        
+
         val baseDir = project.basePath ?: System.getProperty("user.home")
         logger.info("Launching Codex in directory: $baseDir")
-
-        val terminalManager = TerminalToolWindowManager.getInstance(project)
-        var existingTerminal = locateCodexTerminal(terminalManager)
 
         try {
             val httpService = ApplicationManager.getApplication().service<HttpTriggerService>()
             val port = httpService.getActualPort()
-            
+
             if (port == 0) {
                 logger.warn("HTTP service port is not available")
                 notify(project, "HTTP service is not properly initialized", NotificationType.WARNING)
@@ -56,44 +44,13 @@ class LaunchCodexAction : AnAction("Launch Codex", "Open a Codex terminal", null
 
             val settings = service<CodexLauncherSettings>()
             val args = settings.getArgs(port)
-            
             val command = buildCommand(args)
 
-            existingTerminal?.let { terminal ->
-                ensureTerminationCallback(terminal.widget, terminal.content)
-                if (isCodexRunning(terminal)) {
-                    logger.info("Focusing active Codex terminal")
-                    focusCodexTerminal(project, terminalManager, terminal)
-                    return
-                }
+            val terminalManager = project.service<CodexTerminalManager>()
+            terminalManager.launch(baseDir, command)
 
-                if (reuseCodexTerminal(terminal, command)) {
-                    logger.info("Reused existing Codex terminal for new Codex run")
-                    focusCodexTerminal(project, terminalManager, terminal)
-                    return
-                } else {
-                    clearCodexMetadata(terminalManager, terminal.widget)
-                    existingTerminal = null
-                }
-            }
-
-            var widget: TerminalWidget? = null
-            try {
-                widget = terminalManager.createShellWidget(baseDir, "Codex", true, true)
-                val content = markCodexTerminal(terminalManager, widget)
-                if (!sendCommandToTerminal(widget, content, command)) {
-                    throw IllegalStateException("Failed to execute Codex command")
-                }
-                if (content != null) {
-                    focusCodexTerminal(project, terminalManager, CodexTerminal(widget, content))
-                }
-            } catch (sendError: Throwable) {
-                widget?.let { clearCodexMetadata(terminalManager, it) }
-                throw sendError
-            }
-            
             logger.info("Codex command executed successfully: $command")
-            
+
         } catch (t: Throwable) {
             logger.error("Failed to launch Codex", t)
             notify(project, "Failed to launch Codex: ${t.message}", NotificationType.ERROR)
@@ -108,147 +65,6 @@ class LaunchCodexAction : AnAction("Launch Codex", "Open a Codex terminal", null
                 append(args)
             }
         }
-    }
-
-    private fun locateCodexTerminal(manager: TerminalToolWindowManager): CodexTerminal? = try {
-        manager.terminalWidgets.asSequence().mapNotNull { widget ->
-            val content = manager.getContainer(widget)?.content ?: return@mapNotNull null
-            val isCodex = content.getUserData(CODEX_TERMINAL_KEY) == true || content.displayName == "Codex"
-            if (!isCodex) {
-                return@mapNotNull null
-            }
-            CodexTerminal(widget, content)
-        }.firstOrNull()
-    } catch (t: Throwable) {
-        logger.warn("Failed to inspect existing terminal widgets", t)
-        null
-    }
-
-    private fun focusCodexTerminal(
-        project: Project,
-        manager: TerminalToolWindowManager,
-        terminal: CodexTerminal
-    ) {
-        ApplicationManager.getApplication().invokeLater {
-            if (project.isDisposed) {
-                return@invokeLater
-            }
-
-            try {
-                val toolWindow = resolveTerminalToolWindow(project, manager)
-                if (toolWindow == null) {
-                    logger.warn("Terminal tool window is not available for focusing Codex")
-                    return@invokeLater
-                }
-
-                val contentManager = toolWindow.contentManager
-                if (contentManager.selectedContent != terminal.content) {
-                    contentManager.setSelectedContent(terminal.content, true)
-                }
-
-                toolWindow.activate({
-                    try {
-                        terminal.widget.requestFocus()
-                    } catch (focusError: Throwable) {
-                        logger.warn("Failed to request focus for Codex terminal", focusError)
-                    }
-                }, true)
-            } catch (focusError: Throwable) {
-                logger.warn("Failed to focus existing Codex terminal", focusError)
-            }
-        }
-    }
-
-    private fun resolveTerminalToolWindow(
-        project: Project,
-        manager: TerminalToolWindowManager
-    ) = manager.getToolWindow()
-        ?: ToolWindowManager.getInstance(project)
-            .getToolWindow(TerminalToolWindowFactory.TOOL_WINDOW_ID)
-
-    private fun markCodexTerminal(manager: TerminalToolWindowManager, widget: TerminalWidget): Content? {
-        return try {
-            manager.getContainer(widget)?.content?.also { content ->
-                content.putUserData(CODEX_TERMINAL_KEY, true)
-                setCodexRunning(content, false)
-                ensureTerminationCallback(widget, content)
-                content.displayName = "Codex"
-            }
-        } catch (t: Throwable) {
-            logger.warn("Failed to tag Codex terminal metadata", t)
-            null
-        }
-    }
-
-    private fun clearCodexMetadata(manager: TerminalToolWindowManager, widget: TerminalWidget) {
-        try {
-            manager.getContainer(widget)?.content?.let { content ->
-                clearCodexMetadata(content)
-            }
-        } catch (t: Throwable) {
-            logger.warn("Failed to clear Codex terminal metadata", t)
-        }
-    }
-
-    private fun clearCodexMetadata(content: Content) {
-        content.putUserData(CODEX_TERMINAL_KEY, null)
-        content.putUserData(CODEX_TERMINAL_RUNNING_KEY, null)
-        content.putUserData(CODEX_TERMINAL_CALLBACK_KEY, null)
-    }
-
-    private fun reuseCodexTerminal(
-        terminal: CodexTerminal,
-        command: String
-    ): Boolean {
-        ensureTerminationCallback(terminal.widget, terminal.content)
-        return sendCommandToTerminal(terminal.widget, terminal.content, command)
-    }
-
-    private fun sendCommandToTerminal(
-        widget: TerminalWidget,
-        content: Content?,
-        command: String
-    ): Boolean {
-        return try {
-            widget.sendCommandToExecute(command)
-            setCodexRunning(content, true)
-            true
-        } catch (t: Throwable) {
-            logger.warn("Failed to execute Codex command", t)
-            setCodexRunning(content, false)
-            false
-        }
-    }
-
-    private fun isCodexRunning(terminal: CodexTerminal): Boolean {
-        val liveState = invokeIsCommandRunning(terminal.widget)
-        if (liveState != null) {
-            setCodexRunning(terminal.content, liveState)
-            return liveState
-        }
-        return terminal.content.getUserData(CODEX_TERMINAL_RUNNING_KEY) ?: false
-    }
-
-    private fun setCodexRunning(content: Content?, running: Boolean) {
-        content?.putUserData(CODEX_TERMINAL_RUNNING_KEY, running)
-    }
-
-    private fun ensureTerminationCallback(widget: TerminalWidget, content: Content?) {
-        if (content == null) return
-        if (content.getUserData(CODEX_TERMINAL_CALLBACK_KEY) == true) return
-        try {
-            widget.addTerminationCallback({ setCodexRunning(content, false) }, content)
-            content.putUserData(CODEX_TERMINAL_CALLBACK_KEY, true)
-        } catch (t: Throwable) {
-            logger.warn("Failed to register termination callback", t)
-        }
-    }
-
-    private fun invokeIsCommandRunning(widget: TerminalWidget): Boolean? {
-        return runCatching {
-            val method = widget.javaClass.methods.firstOrNull { it.name == "isCommandRunning" && it.parameterCount == 0 }
-            method?.apply { isAccessible = true }?.invoke(widget) as? Boolean
-        }.getOrNull()
     }
 
     private fun notify(project: Project, content: String, type: NotificationType) {
