+++ src/main/kotlin/com/github/x0x0b/codexlauncher/settings/options/ModelReasoningEffort.kt
@@ -1,4 +1,4 @@
-package com.github.x0x0b.codexlauncher.settings
+package com.github.x0x0b.codexlauncher.settings.options
 
 enum class ModelReasoningEffort {
     DEFAULT,
