+++ src/main/resources/META-INF/plugin.xml
@@ -58,7 +58,7 @@
        Read more: https://plugins.jetbrains.com/docs/intellij/plugin-extension-points.html -->
   <extensions defaultExtensionNs="com.intellij">
     <notificationGroup id="CodexLauncher" displayType="BALLOON" isLogByDefault="true"/>
-    <applicationConfigurable instance="com.github.x0x0b.codexlauncher.settings.CodexLauncherConfigurable" id="com.github.x0x0b.codexlauncher.settings" displayName="Codex Launcher"/>
+    <applicationConfigurable instance="com.github.x0x0b.codexlauncher.settings.ui.CodexLauncherConfigurable" id="com.github.x0x0b.codexlauncher.settings" displayName="Codex Launcher"/>
     <postStartupActivity implementation="com.github.x0x0b.codexlauncher.startup.CodexStartupActivity"/>
   </extensions>
